requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
zipp==3.10.0
